+++ prp/parse/mykrobe.py
@@ -2,9 +2,20 @@
 
 import logging
 import re
-from typing import Any, Union
+from pathlib import Path
+from typing import Any, Dict, Union
 
-from ...models.phenotype import (
+import numpy as np
+import pandas as pd
+
+from prp.models.species import (
+    MykrobeSpeciesPrediction,
+    SppMethodIndex,
+    SppPredictionSoftware,
+)
+
+from ..models.metadata import SoupType, SoupVersion
+from ..models.phenotype import (
     AMRMethodIndex,
     AnnotationType,
     ElementType,
@@ -12,9 +23,11 @@ from ...models.phenotype import (
     MykrobeVariant,
     PhenotypeInfo,
 )
-from ...models.phenotype import PredictionSoftware as Software
-from ...models.phenotype import VariantSubType, VariantType
-from ..utils import get_nt_change, is_prediction_result_empty
+from ..models.phenotype import PredictionSoftware as Software
+from ..models.phenotype import VariantSubType, VariantType
+from ..models.sample import MethodIndex
+from ..models.typing import ResultLineageBase, TypingMethod
+from .utils import get_nt_change, is_prediction_result_empty
 
 LOG = logging.getLogger(__name__)
 
@@ -160,13 +173,49 @@ def _parse_mykrobe_amr_variants(mykrobe_result) -> tuple[MykrobeVariant, ...]:
     return variants
 
 
-def parse_mykrobe_amr_pred(prediction: dict[str, Any]) -> AMRMethodIndex | None:
+def _read_result(result_path: str) -> Dict[str, Any]:
+    """Read Mykrobe result file."""
+    pred_res = pd.read_csv(result_path, quotechar='"')
+    pred_res = (
+        pred_res.rename(
+            columns={pred_res.columns[3]: "variants", pred_res.columns[4]: "genes"}
+        )
+        .replace(["NA", np.nan], None)
+        .to_dict(orient="records")
+    )
+    return pred_res
+
+
+def get_version(result_path) -> SoupVersion:
+    """Get version of Mykrobe from result."""
+    LOG.debug("Get Mykrobe version")
+    pred_res = _read_result(result_path)
+    return SoupVersion(
+        name="mykrobe-predictor",
+        version=pred_res[0]["mykrobe_version"],
+        type=SoupType.DB,
+    )
+
+
+def parse_amr_pred(
+    result_path: str | Path, sample_id: str | None = None
+) -> AMRMethodIndex | None:
     """Parse mykrobe resistance prediction results."""
     LOG.info("Parsing mykrobe prediction")
+    pred_res = _read_result(result_path)
+    # verify that sample id is in prediction result
+    if sample_id is not None:
+        if not sample_id in pred_res[0]["sample"]:
+            LOG.warning(
+                "Sample id %s is not in Mykrobe result, possible sample mixup",
+                sample_id,
+            )
+            raise ValueError("Sample id is not in Mykrobe result.")
+
     resistance = ElementTypeResult(
-        phenotypes=_get_mykrobe_amr_sr_profie(prediction),
+        phenotypes=_get_mykrobe_amr_sr_profie(pred_res),
         genes=[],
-        variants=_parse_mykrobe_amr_variants(prediction),
+        variants=_parse_mykrobe_amr_variants(pred_res),
     )
 
     # verify prediction result
@@ -177,3 +226,34 @@ def parse_mykrobe_amr_pred(prediction: dict[str, Any]) -> AMRMethodIndex | None:
             type=ElementType.AMR, software=Software.MYKROBE, result=resistance
         )
     return result
+
+
+def parse_spp_pred(result_path: str | Path) -> SppMethodIndex:
+    """Get species prediction result from Mykrobe."""
+    LOG.info("Parsing Mykrobe spp result.")
+    pred_res = _read_result(result_path)
+    spp_pred = MykrobeSpeciesPrediction(
+        scientific_name=pred_res[0]["species"].replace("_", " "),
+        taxonomy_id=None,
+        phylogenetic_group=pred_res[0]["phylo_group"].replace("_", " "),
+        phylogenetic_group_coverage=pred_res[0]["phylo_group_per_covg"],
+        species_coverage=pred_res[0]["species_per_covg"],
+    )
+    return SppMethodIndex(software=SppPredictionSoftware.MYKROBE, result=[spp_pred])
+
+
+def parse_lineage_pred(result_path: str | Path) -> MethodIndex | None:
+    """Parse mykrobe results for lineage object."""
+    LOG.info("Parsing lineage results")
+    pred_res = _read_result(result_path)
+    if pred_res:
+        lineage = pred_res[0]["lineage"]
+        # cast to lineage object
+        result_obj = ResultLineageBase(
+            main_lineage=lineage.split(".")[0],
+            sublineage=lineage,
+        )
+        return MethodIndex(
+            type=TypingMethod.LINEAGE, software=Software.MYKROBE, result=result_obj
+        )
+    return None