+++ prp/parse/qc.py
@@ -1,4 +1,5 @@
 """Parse output of QC tools."""
+
 import csv
 import json
 import logging
@@ -280,20 +281,24 @@ def parse_postalignqc_results(postalignqc_fpath: str) -> QcMethodIndex:
     with open(postalignqc_fpath, "r", encoding="utf-8") as jsonfile:
         qc_dict = json.load(jsonfile)
         qc_res = PostAlignQcResult(
-            ins_size=None
-            if "ins_size" not in qc_dict
-            else int(float(qc_dict["ins_size"])),
-            ins_size_dev=None
-            if "ins_size_dev" not in qc_dict
-            else int(float(qc_dict["ins_size_dev"])),
+            ins_size=(
+                None if "ins_size" not in qc_dict else int(float(qc_dict["ins_size"]))
+            ),
+            ins_size_dev=(
+                None
+                if "ins_size_dev" not in qc_dict
+                else int(float(qc_dict["ins_size_dev"]))
+            ),
             mean_cov=int(qc_dict["mean_cov"]),
             pct_above_x=qc_dict["pct_above_x"],
             n_reads=int(qc_dict["n_reads"]),
             n_mapped_reads=int(qc_dict["n_mapped_reads"]),
             n_read_pairs=int(qc_dict["n_read_pairs"]),
-            coverage_uniformity=float(qc_dict["coverage_uniformity"])
-            if qc_dict.get("coverage_uniformity") is not None
-            else None,
+            coverage_uniformity=(
+                float(qc_dict["coverage_uniformity"])
+                if qc_dict.get("coverage_uniformity") is not None
+                else None
+            ),
             quartile1=float(qc_dict["quartile1"]),
             median_cov=float(qc_dict["median_cov"]),
             quartile3=float(qc_dict["quartile3"]),