+++ prp/parse/variant.py
@@ -6,7 +6,6 @@ import re
 from cyvcf2 import VCF, Variant
 
 from prp.models.phenotype import VariantBase, VariantType
-from prp.models.phenotype import TbProfilerVariant, VariantBase, VariantType
 
 LOG = logging.getLogger(__name__)
 SOURCE_PATTERN = r"##source=(.+)\n"
@@ -14,22 +13,18 @@ SOURCE_PATTERN = r"##source=(.+)\n"
 
 def _filter_variants(variant_list):
     # Initialize the results dictionary
-    filetered_variants = {
-        'sv_variants': [],
-        'indel_variants': [],
-        'snv_variants': []
-    }
+    filetered_variants = {"sv_variants": [], "indel_variants": [], "snv_variants": []}
 
     # Iterate through each variant in the list
     for variant in variant_list:
-        variant_type = dict(variant).get('variant_type')  # Extract the variant_type
+        variant_type = dict(variant).get("variant_type")  # Extract the variant_type
         # Append the variant to the appropriate key in the results dictionary
         if variant_type == "SV":
-            filetered_variants['sv_variants'].append(variant)
+            filetered_variants["sv_variants"].append(variant)
         elif variant_type == "INDEL":
-            filetered_variants['indel_variants'].append(variant)
+            filetered_variants["indel_variants"].append(variant)
         elif variant_type == "SNV":
-            filetered_variants['snv_variants'].append(variant)
+            filetered_variants["snv_variants"].append(variant)
     return filetered_variants
 
 
@@ -49,13 +44,15 @@ def _get_variant_type(variant) -> VariantType:
 
 def _get_variant_subtype(ref_base, alt_base):
     # Define purines and pyrimidines
-    purines = {'A', 'G'}
-    pyrimidines = {'C', 'T'}
-    
+    purines = {"A", "G"}
+    pyrimidines = {"C", "T"}
+
     # Check for transition substitution
-    if (ref_base in purines and alt_base in purines) or (ref_base in pyrimidines and alt_base in pyrimidines):
+    if (ref_base in purines and alt_base in purines) or (
+        ref_base in pyrimidines and alt_base in pyrimidines
+    ):
         return "TS"
-    
+
     # If it's not a transition, it must be a transversion
     return "TV"
 
@@ -75,11 +72,9 @@ def parse_variant(variant: Variant, var_id: int, caller: str | None = None):
     var_type: VariantType = _get_variant_type(variant)
 
     for alt_idx, alt_var in enumerate(variant.ALT):
-        possible_minority_var = False
         var_subtype = variant.var_subtype.upper()
         if var_subtype == "UNKNOWN":
             var_subtype = _get_variant_subtype(variant.REF, alt_var)
-            possible_minority_var = True
         var_obj = VariantBase(
             id=var_id,
             variant_type=var_type,
@@ -89,8 +84,16 @@ def parse_variant(variant: Variant, var_id: int, caller: str | None = None):
             end=variant.end,
             ref_nt=variant.REF,
             alt_nt=alt_var,
-            frequency=variant.INFO.get("AF") if type(variant.INFO.get("AF")) != tuple else variant.INFO.get("AF")[alt_idx],
-            depth=variant.INFO.get("DP") if type(variant.INFO.get("DP")) != tuple else variant.INFO.get("DP")[alt_idx],
+            frequency=(
+                variant.INFO.get("AF")
+                if not isinstance(variant.INFO.get("AF"), tuple)
+                else variant.INFO.get("AF")[alt_idx]
+            ),
+            depth=(
+                variant.INFO.get("DP")
+                if not isinstance(variant.INFO.get("DP"), tuple)
+                else variant.INFO.get("DP")[alt_idx]
+            ),
             method=variant.INFO.get("SVMETHOD", caller),
             confidence=variant.QUAL,
             passed_qc=passed_qc,