+++ tests/fixtures/saureus/__init__.py
@@ -6,6 +6,12 @@ from ..fixtures import data_path
 
 
 @pytest.fixture()
+def saureus_sample_conf_path(data_path):
+    """Get path for saureus sample config file"""
+    return str(data_path.joinpath("saureus", "sample_1.cnf.yml"))
+
+
+@pytest.fixture()
 def saureus_analysis_meta_path(data_path):
     """Get path for saureus meta file"""
     return str(data_path.joinpath("saureus", "analysis_meta.json"))